# Faster Lua code