@niklas-heer 2022-10-21
Congratulations, it worked for both Lua and LuaJIT 🎉 
![combined_results](https://user-images.githubusercontent.com/1914063/197129199-bc8694d6-fcaa-4503-8b99-35879fe3a682.png)